refactor: menuGroup 패키지 분리
