+++ src/main/java/kitchenpos/menu_group/application/MenuGroupService.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu_group.application;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.repository.MenuGroupRepository;
+import kitchenpos.menu_group.domain.MenuGroup;
+import kitchenpos.menu_group.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
